+++ packages/domain/src/generation/index.ts
@@ -1,6 +1,7 @@
 export * from "./types";
 export * from "./state-machine";
 export * from "./request-identity";
+export * from "./execution-input";
 export * from "./ports";
 export * from "./capability";
 export * from "./prompt-render";